Add a functional /cancel command.
It cancels the current planning and remove it from the database. In order to do so we added the pl_id column to the plannings table in the database and updated all the program logic to handle it correctly. We also prevented any way to create more than one planing at once.